Renaming files to class neames and move some classes and interfaces to they own files.
