+++ src/Prise/Activation/IBootstrapperServiceProvider.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Prise.Activation
+{
+    public interface IBootstrapperServiceProvider : IDisposable
+    {
+        object GetHostService(Type type);
+    }
+}
