+++ src/Prise/Activation/PluginService.cs
@@ -1,15 +1,7 @@
-using System;
 using Prise.Plugin;
 
 namespace Prise.Activation
 {
-    public class BootstrapperService
-    {
-        public string FieldName { get; set; }
-        public Type ServiceType { get; set; }
-        public Type ProxyType { get; set; }
-    }
-
     public class PluginService :  BootstrapperService
     {
         public ProvidedBy ProvidedBy { get; set; }
