+++ src/Prise/AssemblyLoading/DefaultNativeAssemblyUnloader.cs
@@ -1,14 +1,8 @@
 using System;
-using System.Runtime.InteropServices;
 
 namespace Prise.AssemblyLoading
 {
 #if !SUPPORTS_NATIVE_UNLOADING
-    internal static class NativeAssemblyUnloader
-    {
-        [DllImport("kernel32", SetLastError = true)]
-        internal static extern bool FreeLibrary(IntPtr hModule);
-    }
 #endif
 
     public class DefaultNativeAssemblyUnloader : INativeAssemblyUnloader
