+++ src/Prise/AssemblyLoading/IPluginDependencyContextProvider.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Prise.AssemblyLoading
+{
+    public interface IPluginDependencyContextProvider
+    {
+        Task<IPluginDependencyContext> FromPluginLoadContext(IPluginLoadContext pluginLoadContext);
+    }
+}
