+++ src/Prise/Platform/DefaultPlatformAbstraction.cs
@@ -1,0 +1,11 @@
+namespace Prise.Platform
+{
+    public class DefaultPlatformAbstraction : IPlatformAbstraction
+    {
+        public bool IsLinux() => System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Linux);
+
+        public bool IsOSX() => System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.OSX);
+
+        public bool IsWindows() => System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows);
+    }
+}
